Move PathContainer & PathProperty into their own file
